requirements.txt
snips-nlu>=0.20.0,<0.21
dialogflow>=0.7.2,<0.8
docker>=4.1.0,<5
requests>=2.22,<3
matplotlib>=3.1.2,<4
pyyaml>=5.1.2,<6
azure-cognitiveservices-language-luis>=0.5.0,<0.6
langcodes>=1.4.1,<2